+++ app/src/main/java/com/android/notificationexpo/ItemDetailActivity.kt
@@ -11,12 +11,10 @@ import androidx.appcompat.app.AppCompatActivity
 import android.view.MenuItem
 import com.android.notificationexpo.receivers.AlarmManagerReceiver
 
-/**
- * An activity representing a single Item detail screen. This
- * activity is only used on narrow width devices. On tablet-size devices,
- * item details are presented side-by-side with a list of ITEMS
- * in a [ItemListActivity].
- */
+/*Activity che mostra il dettaglio di una specifica chat. Questa classe viene utilizzata solo se il dispositivo utilizzato è
+* uno smartphone. L'interfaccia utente è su un Fragment che viene caricato nel metodo onCreate. Nei tablet il dettaglio di
+* una chat è visualizzato a lato della lista delle chat disponibili.*/
+
 class ItemDetailActivity : AppCompatActivity() {
 
     // Broadcast receiver
@@ -42,24 +40,18 @@ class ItemDetailActivity : AppCompatActivity() {
         //mostro la action bar
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
-        // savedInstanceState is non-null when there is fragment state
-        // saved from previous configurations of this activity
-        // (e.g. when rotating the screen from portrait to landscape).
-        // In this case, the fragment will automatically be re-added
-        // to its container so we don"t need to manually add it.
-        // For more information, see the Fragments API guide at:
-        //
-        // http://developer.android.com/guide/components/fragments.html
-        //
+        /*savedInstanceState è non null quando c'è un fragment già creato da una configurazione precedente di ItemDetailActivity
+        * (per esempio quando si ruota il dispositivo). In questo caso il fragment verrà automaticamente riaggiunto nel
+        *  FrameLayout (con id item_detail_container) perciò non è nessario aggiungerlo manualmente*/
         if (savedInstanceState == null) {
-            // Create the detail fragment and add it to the activity
-            // using a fragment transaction.
+            // Creo il fragment del dettaglio e lo aggiungo all'Activity
+            // usando una fragment transaction.
             val fragment = ItemDetailFragment().apply {
                 arguments = Bundle().apply {
                     //passo al ItemDetailFragment le informazioni di cui ha bisogno
                     putLong(ItemDetailFragment.CHAT_ID, intent.getLongExtra(ItemDetailFragment.CHAT_ID,-1))
                     putString(ItemDetailFragment.CHAT_NAME, intent.getStringExtra(ItemDetailFragment.CHAT_NAME))
-                    putInt(ItemDetailFragment.CHAT_IMG,intent.getIntExtra(ItemDetailFragment.CHAT_IMG, -1) as Int)
+                    putInt(ItemDetailFragment.CHAT_IMG, intent.getIntExtra(ItemDetailFragment.CHAT_IMG, -1))
                     putString(ItemDetailFragment.NOTIFICATION,intent.getStringExtra(ItemDetailFragment.NOTIFICATION))
                     putBoolean(ItemDetailFragment.TWO_PANE,intent.getBooleanExtra(ItemDetailFragment.TWO_PANE,false))
                 }
@@ -74,12 +66,7 @@ class ItemDetailActivity : AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem) =
             when (item.itemId) {
                 android.R.id.home -> {
-                    // This ID represents the Home or Up button. In the case of this
-                    // activity, the Up button is shown. For
-                    // more details, see the Navigation pattern on Android Design:
-                    //
-                    // http://developer.android.com/design/patterns/navigation.html#up-vs-back
-
+                    //Questo ID rappresenta il pulsante Home
                     navigateUpTo(Intent(this, ItemListActivity::class.java))
                     true
                 }
